Extracted ReflectionCache in order to use it in JsonSchemaReferenceUtilities.
