+++ src/NJsonSchema/NJsonSchema.csproj
@@ -90,6 +90,7 @@
     <Compile Include="JsonSchemaReferenceUtilities.cs" />
     <Compile Include="JsonSchema4.Serialization.cs" />
     <Compile Include="JsonSchema4.cs" />
+    <Compile Include="Infrastructure\ReflectionCache.cs" />
     <Compile Include="Validation\JsonSchemaValidator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Generation\JsonSchemaGenerator.cs" />
